requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attr==0.3.2
attrs==25.3.0
Authlib==1.3.2
billiard==4.2.1
blinker==1.9.0
celery==5.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.0
distro==1.9.0
dj-database-url==3.0.1
Django==5.1.2
django-allauth==65.10.0
djangorestframework==3.15.2
drf-yasg==1.21.10
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
flask-oidc==2.2.2
fsspec==2025.3.0
gunicorn==23.0.0
h11==0.14.0
has==0.0.4
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
image==1.5.33
inflection==0.5.1
instaloader==4.14.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.9.0
kombu==5.5.3
lxml==5.3.1
MarkupSafe==3.0.2
numpy==2.1.2
openai==1.66.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.51
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytrends==4.9.2
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
selenium==4.29.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.49.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.2.3
vine==5.1.0
waitress==3.0.2
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
whitenoise==6.9.0
wsproto==1.2.0
